fix(server): fix tests due to Swagger SNAPSHOT update
Swagger changed the way it serializes the Java model, now it includes
`responseSchema` property. We test against the Swagger specification
embedded as a part of the generated connector so we have to adapt.

Wonder where that `responseSchema` property comes from it's not in the
specification. We'll probably need to revert this or modify it to make
it work with another -SNAPSHOT version.

Living on the edge.