fix(ui): make icons reflect changes
As the URL of the icon never changes when modifying the icons of API
connectors the old icon is not replaced with the newly uploaded one
regardless of `Cache-Control` headers in Firefox.

This adds the icon id, from the database, as a query parameter to the
icon URL which will make it reload the icon when it's changed.

Also adds `iconFile` property as it is set at first icon submission in
on the `CustomConnectorRequest`. This makes the icon change show up
immediately. Very quickly after that the icon URL will be replaced from
the data URL to the URL on the server by the event received from the
server that the connector has changed.

Fixes #1433