chore(server): implement global transaction with JsonDB
This adds global transaction support to JsonDB, that is a way of
invoking methods on JsonDB multiple times including the methods that use
transaction support, but still within a all-or-nothing transaction.

In the SqlJsonDB implementation this is done by instantiating a new
instance of SqlJsonDB and providing it with a single connection that has
transaction methods ignored. To be fully compliant with JsonDB contract
transacted event bus is also added so events will be propagated only in
case of the successful transaction execution.