refactor(server): add server-runtime jar with classes classifier
This adds a `server-runtime` JAR with `classes` classifier that replaces
the previous no-classifier plain (non Spring Boot) JAR we used as
dependency. This `classes` classifier JAR also excludes
`application.yml` and several other resource files that in general are
not needed when depending on the `server-runtime` module as a Java
dependency.