fix(server): ignore "duplicate" application.yml
The duplicate-finder-maven-plugin finds a duplicate resource
`application.yml` in `target/classes` of the `server-runtime` module and
reports a duplicate classpath resource issue.

The `application.yaml` is not present in the `syndesis-server`/`classes`
classifier JAR file used and bundled in the fat JAR, so this adds the
resource to `ignoredResourcePatterns` to pass the check.

Fixes #2164