fix: slimmer inspections in connectors
Seems that Atlasmap inspections embedded in connector JSON files are
causing the the Connection/Connector objects to be quite large. This is
mainly due to the use of pretty-printing when generating those embedded
JSON specifications.

By not using a pretty-printing JSON serialization this can be greatly
reduced, for 5 test connections the JSON payload for is reduced from
1.2M (1248066) to 685K (700638), i.e. 44% decrease in size.