feat: implement refresh of access token
This implements additional logic and configuration options for the API
connector functionality.

OAuth2 access token refresh functionality is added to REST Swagger
connector. Circumstances under which this functionality is enabled is
the folowing:
 - API is using OAuth2 authentication type
 - refreshToken property was given, i.e. refreshToken was issued and
   received when the Syndesis connection was created in the OAuth2 flow
 - there is at least one HTTP status defined for which the refresh of
   the access token is performed

When the Camel exchange signals a HTTP error via
HttpOperationFailedException, this logic kicks in and if the list of
HTTP status codes contains the status code of the received response and
previous attempt (if any) to refresh the access token was not with the
same refresh token an attempt is made to to refresh the access token
with the current refresh token.

If we receive a new refresh token during the refresh of the access token
we will remember it for the next access token refresh in the integration
pod, the component property is mutated.

The first exchange that fails due to the expired access token will fail,
i.e. this will not retry that exchange if we receive a new valid access
token.

What the user needs to perform in order to trigger this functionality:
 - the user needs to add a vendor extension
   `x-refresh-token-retry-statuses` to the `securityDefinitions` of the
   OAuth2 security definition in the Swagger specification -- without it
   this functionality will not be performed

Example:
```yaml
securityDefinitions:
  oauth2:
    type: 'oauth2'
    flow: 'accessCode'
    authorizationUrl: '...'
    tokenUrl: '...'
    scopes:
      ...
    x-refresh-token-retry-statuses: 403
```

This will instruct the connector to refresh the access token for HTTP
403 status codes.

Fixes #2408