refactor(connector): edge case and moar tests
This fixes an edge case whereby rest-swagger connector would not refresh
the OAuth access token if the service tends to use the same refresh
token.

The logic of not retrying with the same refresh token in that case is
flawed as it is legitimate to use the same refresh token if the service
upon refresh of the access token returns the same refresh token.

Also cleans up the implementation and adds more unit tests.