fix(server): use proper integration version
When generating OpenShift Secret we used the latest version of the
Integration. That meant that if we modified the order or added steps the
Secret would be misaligned with the image build used. For instance if we
removed a logging step (at position 2) between the two existing steps:
start at positon 1 and stop at position 3 steps, the id of the last step
would changed from 3 to 2, but the generated `application.properties` in
the Secret would still have the key with `-3`.

This prevented deployments of previous versions or deployments of newer
versions.

Fixes #2432