chore(server): set default max-age and expiry on client state cookies
If when creating OAuth connections the user cancels or gets an error the
client side state cookie might remain, this is an issue if many of these
cookies are present in the browser as there are limits on HTTP header
sizes in the browser and the server.

This adds default max-age and expiry to the cookie so it will be removed
from the browser at the same time as it times-out on the server.

Fixes #2443