fix(ui): remove the oauth settings by setting values to null
Syndesis server used to consider properties that are empty or contain
only whitespace as `null` (_trim to null_), this is no longer the case
and the property value needs to be specified as `null`.

Removing OAuth settings would only change the value to an empty string
(`""`) and thus would be still remain as configured.

Fixes #2533