feat(server): toggle `derived` flag based on oauth settings
This adds toggling of the `derived` flag of Connections created for a
Connector whose OAuth client ID and secret have been either removed or
configured. This in turn allows the UI to present the user with the form
containing all properties (derived=false) or hide all properties
(derived=true).

Fixes #2533