fix(server): don't warn about missing properties on derived connections
For Salesforce Connector Connections we support multiple ways of
configuring the properties, when the connection is configured without
OAuth settings set, i.e. manually, the username and password fields are
required; with OAuth settings, i.e. using OAuth flow, username and
password are not set and not required.

Given that the warning SYNDESIS006 is issued if a required property is
not configured, a warning will be shown on Salesforce connections
created from OAuth flow. We can distinguish those by having their
`derived` flag set and filter out this warning.

Fixes #2532