fix(build): use proper parent for documentation
This changes the parent of the documentation POM to the Syndesis parent
and the parent of the documentation maven plugin to the documentation
parent in an effort to help the release build find the documentation
plugin and have all POMs inherit from Syndesis parent.

Also fixes dependency/codestyle issues in the Maven documentation plugin
and makes one of the links relative to the rest of the documentation.