fix: don't allow deleting concur connector
This removes connector template id so that the SAP Concur connector
doesn't show up in the API connectors under Customization, so it cannot
be deleted.

Also sets the `connectorGroupId` to `io.syndesis.connector.concur` and
removes the `actionsSummary` property that's being calculated at runtime
by Syndesis server, so no need to bundle it in the connector definition.

Fixes #2854