fix(server): k8s secret cannot be created
This reduces the size of the `specification` field by removing the
values not consumed by the REST Swagger Camel component and thus reduces
the overall size of the generated k8s secret.

Changes the Concur connector, so that it's `specification` property is
containing only the required (by Camel Component) properties.

Fixes #2880