chore(ui): logout endpoint
This adds a `/logout` endpoint backed by nginx and static logout.html
file.

When `/logout` endpoint is requested with the 'SYNDESIS_XSRF_TOKEN` set
to `awesome` the response will expire the `_oauth_proxy` cookie. In any
case it will serve the `logout.html`.

The UI when performing the logout will replace the `window.document`
with the response of the `/logout` endpoint and will set the URL to
`/logout`.

This design protects the denial of service if the `/logout` endpoint is
invoked cross origin and allows for the served page to be reloaded
making it deterministic and thus safer.

The `logout.html` embeds the used CSS and just displays a simple _You
have been logged out_ message offering to Login again by linking to `/`.

Fixes #2561