chore: OAuth app settings backed by configurati...
...on from the backend

This changes the model of OAuthApp so it no longer has fixed properties.
Now the model is backed by properties/configuredProperties convention we
use for Connectors/Connections/..., allowing the properties as defined
by the Connector.

The properties used/exposed are determined by tag values, and those are
preexisting tags that are already used within the credentials module and
the Connector definitions.

Fixes #2990