fix(server): don't perform updates without changes
When we compare periodically state of objects in controllers we run for
integration deployments, integrations, connections and connectors we
should not generate updates of the JsonDB or emit changes via WebSockets
if the state did not change.

This checks if the previous state is equal to the current state and only
performs the database update end emits the WebSocket event if the change
has occurred. Thusly it increases our performance by reducing the number
of refreshes in the UI and decreasing the database load for updates and
fetches initiated by the emitted change to the UI.

Fixes #3071