fix(connector): Salesforce output unmarshalling
The Salesforce connector was trying to unmarshall the outgoing exchange
message where it should have used the incoming message. This caused
the output of the Salesforce connector for create/create-or-update and
notifications (starting) actions to be `null` instead of the expected
type populated with data from the Salesforce response.

Fixes #2853