feat: mark Connectors provided as tech preview
Adds a new tag `tech-preview` that can be added to connectors that are
provided for early access technology preview.

First such Connector is the IRC Connector.

Visually when Connection is displayed the top of the Connection card
will contain `Technology preview` text, which is also displayed during
connection creation.

Fixes #3201