fix(server): use timeout of 35mins
This adds a read timeout that should provoke `SocketTimeoutException` if
no data was received for 35mins. This is to fix an issue whereby the
client keeps trying to read from a half-open connection and never
receiving any data or any notice that the server has stopped sending the
data.

Fixes #2598 (perhaps)