refactor(build): when installing with --local u...
...se provide the operator with local template

Now when we build with `--local` operator uses the template it was built
with and it is contained in the image of the operator.

This introduces a `syndesis-template` ConfigMap that contains the
content of the local `syndesis.yml` OpenShift template. The
`syndesis-operator` DeploymentConfig is patched to mount this template
from the ConfigMap into the `/conf` directory and to use that template
when creating Syndesis resources.