refactor(db): replace database maintenance Cron...
...Job with aggressive autovacuum configuration

It's difficult for us to use the CronJob as the container image used
would need to be productized. Seems that as CronJob is a non-OpenShift
resource referencing the same PostgreSQL image we use for the database
pod -- referencing via lookup image stream from other namespace (in this
case openshift) is currently not possible.

With this we loose `FULL ANALYSE` from `VACUUM` and `REINDEX TABLE` for
`jsondb` table, the hope is that a more aggressive autovacuum
configuration will be good enough.

See https://dba.stackexchange.com/questions/21068/aggressive-autovacuum-on-postgresql

Fixes #3252