fix(server): drop activity logs if we cannot wr...
...ite to the database

When we are unable to write to the database, for instance when the
database storage runs out of disk space, we should drop activity logs
that are accumulating in the `ActivityTrackingController.eventQueue` as
there is a fixed limit imposed on that queue (1000) and when that limit
is reached no further writing to the database will be done. This in turn
will create a lot of new threads for new requests and leave a lot of
open connections (one every 35min) to the k8s API server.

The solution here doesn't terminate the the processing of the event
queue in case of errors while attempting to write to the database and
the activity logs from that batch are effectively dropped.

This increases our robustness at the expense of consistency.

Fixes #3337