fix(server): thread safety, prevent explosion o...
...f threads

Removes the rescheduling of log watcher, the scheduling already is
performed when pods are polled (every 5 seconds).

Adds `AtomicBoolean` for tracking if the log watcher is running to
prevent multiple schedules for the same pod.

Makes sure that we're tracking log handlers in thread safe manner
by using thread safe collections.