fix(server): when discarding activity logs leav...
...e the newest entries

We were missing descending ordering of the `path` column. The `path`
column contains the key generated when the activity is persisted, the
key is chronologically ordered -- so newer keys are lexically greater
than older keys. That means we should have ordered in descending order
(`ORDER BY path DESC`) instead of the default, which is ascending.

Fixes #3396