fix(ui): logout on firefox
This changes the Nginx configuration to introduce two location blocks
one for the `/` (`/index.html`) and one for all other static resources
under `/`. Seems that `always` option is required for our use case in
the `add_header` directive.

The overall logic is that we can't let the browser to cache `/`, if it
does it will immediately show the Syndesis chrome instead of issuing the
request and this will result in bypassing the `oauth_proxy`
authentication check, i.e. not mandating the user to re-login.

I've found the easiest way to test is to test with a separate Firefox
profile (`firefox --P test`) and in that profile set the History
preference to custom and configure _Clear history when Firefox closes_
option.

Fixes #3016