fix: support Concur APIs with query parameters
With #3564 backported to 1.4.x we fixed the issue with all custom API
client connectors generated by the user in the Custom API connector
wizard. We did not fix the issue with the Concur connector that is
packaged out of the box and has the OpenAPI specification without the
query parameters that were removed in the minimized (see #2880) OpenAPI
specification that we store.

This is a backport of the OpenAPI fixed specification for the Concur
connector contained within the packaged Concur connector descriptor.