chore: add additionalQueryParameters for OAuth
Adds `additionalQueryParameters` map of, well, additional query
parameters that will be passed to the OAuth authorization endpoint when
initiating the OAuth dance.

Values of for the `additionalQueryParameters` map are found in the
connector's configured properties on a property tagged with
`oauth-additional-query-parameters`.

For example:

```json
{
  "id": "sample-connector",
  "name": "Sample Connector",
  "configuredProperties": {
    "additionalOAuthQueryParameters": "{\"access_type\": \"offline\"}"
  },
  "properties": {
    "additionalOAuthQueryParameters": {
      "tags": [
        "oauth-additional-query-parameters"
      ]
    }
  }
}
```

Will add `access_type=offline` query parameter to the authorization URL.