refactor(openshift): selective exposure
This changes the way exposure of integration pods is done by introducing
a choice of exposing the k8s Service and/or OpenShift Route.

Previously we had the choice of either exposing or not exposing the
integration, with this we now have the choice to add more ways of
exposing the integration pods.

The functionality should not be changed by this; it just lays ground
work for integration with 3scale by exposing only k8s Service without
the OpenShift Route.

Corrects the tests so they assert the interaction with the API server by
checking the request payloads, previous version only configured mock
responses without checking the request payloads.