fix(oauth): handle hidden properties defultValue
When we made the change in #3674 (PR #3678) not to send hidden
properties to the UI -- the UI stopped caring about the values of those
properties (it simply wasn't aware of them). So when the OAuth settings
were changed by the user the value of the property was set to `null`
overriding the previously set `configuredProperty`.

This changes the descriptors of gmail and google-calendar connectors so
that the values needed to be set in `configuredProperties` are defined
as `defaultValue` of the property and the logic was implemented to set
the `defaultValue` to the `configuredProperty` if the recived value from
the UI is `null`. Exactly as it functioned when the UI was aware of the
hidden properties.

Also we were missing a tag in the tags to consider when setting
`configuredProperties` of a Connector based on the input from the user
and the Connector metadata in `properties`.