refactor: handling of hidden OAuth connector pr...
...operties

With this we can configure in the Connector metadata hidden
`configuredProperties` or give `defaultValue`s where appropriate. The
change makes the OAuth-tagged `properties` that have `type=hidden` and
for which values from the UI are not received (as the UI has no
knowledge of them, from #3674 (PR #3678)) take values from
`configuredProperties` of the connector or from `defaultValue` of the
`property`, whichever is defined first (in that order).