feat: proper base path handling in API provider...
... integrations

This introduces `properties`/`configuredProperties` on the `Integration`
model. This is needed to populate the `api-basePath` property that holds
the base path of the API provider based integration.

This base path is then passed to the integration pod via Kubernetes
Secret as usual in the generated `application.properties` under the same
key.

This will hold true for all `configuredProperties` on the `Integration`.

The Swagger specification now includes the base path and the
specification is served from `<base-path>/.api-doc/swagger.json`.

3scale integration annotations will have the correct `path` and
`description-path` values

Base path of `/` is not supported.

Fixes #3636