fix: UT010029: Stream is closed with WebHook co...
...nnector

This replaces the message on the exchange not only the body, so the type
of the message is no longer `HttpMessage` and when copies of the message
are attempted the HTTP request stream is not read the second time. At
which point the HTTP request stream might be closed.

Fixes #3727