chore(server): response data shapes rework
This changes the data shapes generated for the use in the custom API
client connector and the API provider connector, so that the request and
response data shapes are in line with the Syndesis HTTP data shape.

For example if the previous response data shape defined as:

```json
{
  "a": "b"
}
```

With this change the the data shape will be defined as:

```json
{
  "parameters": {
    "Status": 123,
    "Content-Type": "x/y"
  },
  "body": {
    "a": "b"
  }
}
```

This is just a example, the data shape is defined as JSON Schema.