chore(upgrade): Update release script to set the proper upgrade pod tag.
Also removed the SYNDESIS_VERSION parameter for the upgrade template
used by the operator.
The version for this pod will be hardcoded and created when the template is created.
While we could assume operator version == upgrade-pod version in upstream (there everything is in sync),
this is not the case for productisation where build number might differ.

Therefore it has to be ensured already during template creating that the
proper tag (including build number) is set.

The tag for the ugrade pod needs to be provided via --upgrade-tag when
creating the templates.
This has been added for the community release and it shoud be verified
that this option is also used during productisation.

Fixes #3463.