chore: use `http` instead of `https`
When OpenShift is running without a trusted CA issuing the certificates
REST API connector will fail with TLS error as the certificate for the
TODO route will not be valid. This changes the scheme from `https` to
`http` and adds a step to prompt the users to change it on their end
also.