chore: introduce wrapped JSON schema identifier
This introduces the wrapped JSON schema identifier of
`io:syndesis:wrapped`, we use it to detect if the response needs to be
_wrapped_ i.e. convert `{"hello": "world"}` to
`{"body": {"hello: "world"}}`.

Also tries to stay backward compatible. Because of this there is a
possibility to detect some payloads as wrapped, and that's why it's
better to rely on the schema identifier.

Fixes #3914