fix: use HeaderFilterStrategy with all HTTP
Previous attempts of removing Camel message headers did not pan out
successful, this is mainly because it's a race between the Activity
tracing interceptor that adds this to around every interceptor and the
header-removal logic.

With this, a new approach of using/configuring HeaderFilterStrategy
to the servlet (webhook & API provider), HTTP and API client connector
eliminates the need for removing Camel message headers and deals with
the issue on the HTTP protocol binding level.

As the HTTP4 component RestProducerFactory implementation overwrites any
HeaderFilterStrategy configured on the component/endpoint a subclass was
necessary to make this function.

So with this fix all HTTP interactions: be it HTTP requests sent out of
integrations (HTTP connector, API client connector) or any HTTP
responses received from integrations (webhook, API provider) should not
contain `Syndesis.*` HTTP headers regardless of their presence in the
Camel message headers.

Fixes #3865