chore: create and delete OAuthClient via operator
This modifies the Syndesis operator Docker image to include
`syndesis-privileged.yml` OpenShift template instead of `syndesis.yml`,
as the `syndesis-privileged.yml` contains the definition for the
Syndesis OAuthClient.
It also removes OAuth service account it is no longer needed (replaced
by OAuthClient).
Now a new ClusterRole/ClusterRoleBinding is needed.
When the Syndesis CRD is deleted, the corresponding OAuthClient is
removed manually, it is not GCed, possibly because it is not within the
same namespace.