refactor: set the phases on for Syndesis Maven ...
...plugins

This helps with invoking maven goals that do not require Syndesis Maven
plugins (like `mvn clean`) when the said Maven plugins are not available
from the local repository (or for download).

I don't see us hosting a repository with SNAPSHOTS (we'll not at least
in the short-to-medium time), this at least helps invoking some Maven
phases when the Syndesis Maven plugins are not required.

Depending on Maven plugins from the same project/reactor is usually not
a good idea.

Fixes #1501 (sort of)