refactor(server): additional fixes for #4477
We need to make sure that the configuration changes and mixins of
Jackson ObjectMapper used by SwaggerParser's are applied. And the way to
make sure that they are is to make sure that all usage of SwaggerParser
is done through `OpenApiHelper` which was created by refactoring bits
of `SwaggerHelper` from `server-api-generator` into `common-util` as it
is used in `integration-project-generator` also.

Ref #4477