refactor(server): perform usage accounting of c...
...onnections and extensions on integration update

Adds a new `ResourceUpdateHandler` to update the connection usage when
integration is updated and persists the change in the database. This way
an event that the connection is updated is fired and broadcast to the
UI.

Impementation contains a debounce of 3 seconds so that the backend
database is not overwhelmed by updates. Some care has been taken not to
interact with the database too much, though all integrations and all
connections need to be fetched with the current interface.

Fixes #3931