fix: filter out all headers but Content-Type
Previously we only filtered out headers with names containing
`Syndesis`, `Camel` or `org.apache.camel`. This allowed headers received
from outside to be passed on to 3rd parties. To be on the safe side this
filters out all headers except for the `Content-Type` header which we
need internally when dealing with Syndesis HTTP message body containing
both parameters and body (unified data shape).

Fixes #3727