refactor: ui build
Splits two profiles into multiple profiles that can be/are activated
independently, while cleaning up and refactoring property names to be
more explicit.

For example `with-proxy` profile can be configured using `proxy-*`
properties (`-user`, `-password`, `-server`, `-port`) and is activated
if `proxy-server` property is defined.

Custom NPM/yarn registry is configured in `with-custom-registry` profile
if `custom-registry` property is defined.

Non-validation of TLS certificates is triggered by activating the
`insecure` profile (`-Pinsecure`). Same with the `verbose` profile to
introduce debug output.

And `node-sass` is re-fetched and re-compiled if `sass-binary-site`
property is defined.