feat(connector): add support for API key authen...
...tication

This adds support for API key authentication, or the parameter
authentication as we choose to call it in the `rest-swagger` connector.

The contract is this: in the OpenAPI definition if an operation has a
security requirement and the OpenAPI definition's security definitions
contain this definition (keyed by name) and it's type is `apiKey` and
the connector's configured properties contain
`authenticationParameterName=<name of the parameter>`,
`authenticationParameterPlacement=<header|query>` and
`authenticationParameterValue=<the key>` properties, a query or a header
parameter will be added to the HTTP request.