fix(server): instead of `operation-NNN` use ran...
...dom UUID

When editing the OpenAPI document it is possible that the current
operation ID algorithm of using the base of `operation-` and adding the
monolithically increasing number (`operation-NNN`) will generate the
same operation ID of a removed operation and in doing so inheriting that
flow.

This uses random UUID instead so that two identical operation IDs cannot
be generated.

Fixes #5310