fix: DATAVIRT_ENABLED parameter use in sydesis.yml
This fixes the inclusion of DATAVIRT_ENABLED template parameter in the
syndesis.yml template. The value needs to be inserted as a number and
using double curly braces (`${{DATAVIRT_ENABLED}}`) makes sure that this
is done.

For mustache template the delimiter needs to be changed so `{{...}}` is
not interpreted so it's first changed to `((..))` and later changed back
to `{{..}}`.

Fixes #5442