chore: switch the syndesis-ui image to React
This switches the `syndesis/syndesis-ui` image to React and publishes
the Angular UI as `syndesis/syndesis-ui-legacy`. The fabric8 Maven
plugin configuration was changed to reflect that and the CircleCI
configuration was updated to build the two versions as separate jobs.

OAuth proxy configuration was switched to the configuration needed for
the React UI.

Ref #5426