fix(ui): display correct icon for API provider ...
...integrations

Seems that we were for some reason displaying the icon of the first step
of the first flow instead of the `{api}` icon that we display on the
integration details page. This makes it fail when there are no flows
defined (i.e. no operations) in the API provider integration.