fix(build): disable generation of source maps
In a bit of a heavy weighted manner this disables source map creation
in the `syndesis` workspace of the React codebase. This makes the build
take much less (eye measurement ~600MB vs 1.5GB). I think we would wish
to configure this depending on the `NODE_ENV` and perhaps just for the
`production` build, though this should fix the issue we have right now.

Fixes #5468