fix(connector): remove tags from rest-swagger
Custom API client connector had tagged properties, so when loaded into
the database it would be detected as OAuth-capable connector which in
turn would add it to the Settings page of the app.

This removes all the tags as they will be added by the rest-swagger
connector template in the `deployment.json`.

Fixes #5532