fix: don't reset OAuth client state
We need to destroy the state received from backend on API call to
`/api/v1/connectors/{id}/credentials` in order for it not to overwrite
the cookie received on `/api/v1/credentials/callback`. This eliminates
the need to parse the cookie in `oauth-redirect.html` and pass/store it
in the app.