feat(api): add `PUT /api/v1/apis/generator`
When creating an integration from the OpenAPI document prior to saving
the integration there is a need to modify the OpenAPI document and have
the changes reflect in the integration.

This adds a `PUT /api/v1/apis/generator` that returns HTTP status `202`
and the updated integration when the provided OpenAPI document produces
changes to the integration; and HTTP status `304` with no content when
there are no changes to the integration.

There's a side effect of storing the updated OpenAPI document and
referencing it from the integration resources, if there was an update.

Fixes #5719