fix: distinguish between differently named secu...
...rity definitions

When several security definitions of the same type are present in the
OpenAPI document it's difficult to distinguish between them.

This adds name and description (if present) to the label and type
followed by the name separated by colon in the value of the
`authenticationType` property `enum` values. This serves as distinction
for the user and also for the code handling those `enum` values.

Fixes #5669