feat: require OAuth settings for connection cre...
...ation

When creating OAuth-capable connections issues can arise when connection
properties are configured by manual entry. For instance refresh token
that is hidden from the UI will not be filled and the connection will
not be valid as the access token for it will expire and could not be
renewed.

This introduces a change in the connection creation that if a new
OAuth-capable connection creation is attempted OAuth settings must be
set. The user is prompted to enter the properties in OAuth settings and
when those are set, connection creation via OAuth flow is allowed.

Fixes #5582