fix: handle credentials response without transf...
...ormation

When `GET /api/v1/connectors/{id}/credentials` response is `{}` the
`useApiResource` will provide `undefined` as the the response. This
causes an issue as `Object.keys(undefined)` throws an Error. Also there
is no need to transform it to `undefined` as the code that handles the
response handles both `{}` and `undefined` cases as the result from
`useConnectionCredentials` is typed `<AcquisitionMethod | undefined>`.

The check later in the code tries to determine by accessing
`AcquisitionMethod.type` and `AcquisitionMethod.configured`, but also
handles when the returned `AcquisitionMethod` is `undefined`.

Fixes #5831