fix: honor both syntaxes for authenticationType
With #5669 we added support for adding a name of the security definition
to distinguish between two authentication methods of the same type to
the `authenticationType` configured property value. At the same time
when the connector is to be created the authentication method selected
by the end user needs to be paired with the security definition from the
OpenAPI document.

Here we did not account for the additional name of the security
definition that can be present in the value of the `authenticationType`
configured property. This makes sure that we support both variants in
the `authenticationType` configured property.

Fixes #5855