fix: pass `configuredProperties` from connector...
... in OAuth flow

When we create a connection in OAuth flow we need to pass the
`configuredProperties` that have been pre-configured on the connector.

Another way to fix this would be to include `configuredProperties` from
the connector when passing parameters to connection validation. This
makes it in-line for OAuth based connectors as it does for non-OAuth
based connectors.

Fixes #5891