refactor: don't use flow id to correlate API op...
...erations

Before this we correlated API operations from OpenAPI document by
embedding the integration id and the operation id in the flow id. This
stores the operation id in the flow metadata and uses key generator
generated keys for flow ids.

List of endpoint changes:

 - `POST /api/v1/apis/generator` doesn't set integration id, flow ids
   are key-generated, operation ids are placed in metadata
 - `PUT /api/v1/apis/generator` and
   `PUT /api/v1/integrations/{id}/specification` correlate using
    operation id stored in the metadata

Schema was upgrated to 30 and a migration was added to change the flow
ids and put the operation id in the flow metadata.

Fixes #5963