fix: don't require authentication for /config.json
Seems that the request for `/config.json` is resulting in a 302 redirect
to the HTML page served from oauth-proxy to authenticate. Parsing this
response as JSON fails and we get an error in the UI.

This only seems to happen on a particular installation, and I think this
should fix this issue. Having public `/config.json` doesn't seem like a
security issue to me.

Fixes #5602