fix: AcquisitionMethod JSON serialization
With #5924 the `configured` was refactored to `isConfigured` so that the
Swagger core JAX-RS support recognizes the property. This made the
serialization to JSON serialize it as `isConfigured` instead of
`configured` as declared in the internal OpenAPI document.

This reverts that refactor, and the adds `@JsonProperty` annotation to
explicitly name the property as `configured`.

Fixes #6005