chore(tool): configure devImages: true
This configures Syndesis custom resource option `devImages` to `true`,
to configure that the images for meta, server and ui are not pulled
from Docker Hub but from the internal OpenShift registry.

When running `syndesis minishift --install` or `syndesis install --dev`
this will result in no meta, server or ui pods being deployed as they
first need to be built.