fix(atlasmap-assembly): remove routing related console warning
Since Atlasmap doesn't use the routing, we can avoid enabling it. This
gets rid of the warning we were having about the router not being happy
about being initialized in an iframe without an src attribute.

Also, fixes the minishift setup/restore script to work with the
operator. Since the operator reverts back any local change to operator's
managed resources, we have to scale it down to 0 when running the setup
script. The restore script brings the operator back to 1 replica.